requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.11.0
anthropic==0.42.0
voyageai==0.3.2
stripe==11.4.1
python-dotenv==1.0.1
pydantic==2.10.4
tiktoken==0.8.0
